requirements.txt
# Web Framework
Flask==3.0.3
Werkzeug==3.0.3

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
PyMySQL==1.1.1
cryptography==42.0.7
Flask-Migrate==4.0.7

# CORS
Flask-Cors==4.0.1

# Environment / Config
python-dotenv==1.0.1

# Validation
marshmallow==3.21.3

# Video streaming (MJPEG — headless: no GUI deps)
opencv-python-headless==4.9.0.80

# Deep Learning / Computer Vision
ultralytics>=8.2.0
Pillow>=10.3.0
PyYAML>=6.0.1
tensorflow>=2.16.0

# Production WSGI server
gunicorn==22.0.0

# Dev / Test
pytest==8.2.2
pytest-flask==1.3.0
